+++ src/shared/routes.ts
@@ -1,0 +1,5 @@
+export enum PageLinks {
+  MAIN = "/",
+  PAYMENT = "/payment",
+  DASHBOARD = "/dashboard",
+}
